Changed to size_t & added dbg output.

+++ net_core/include/node.h
@@ -13,7 +13,7 @@
 /**
  * @brief Add a message to a queue so that it can be procecedd later.
  */
-void node_enqueue(uint8_t *data, uint8_t length);
+void node_enqueue(uint8_t *data, size_t length);
 
 /**
  * @brief Send a messsage over the radio with an added header.
